pipe

+++ stop.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
-#epmd -names
 HN=`hostname`
-echo `ps auxwww | grep "goanna@$HN"`
 PID=`ps auxwww | grep "goanna" | grep -v grep | awk '{ print $2 }'`
-echo $PID
-kill $PID
+if [ ! -z "$PID" ]; then
+    set -x
+    kill $PID
+fi
